+++ src/client/pages/forgot-password/index.jsx
@@ -95,6 +95,7 @@ export default class ForgotPassword extends PureComponent {
           <Button
             className={styles.submitBtn}
             action="submit"
+            type="submit"
             primary
             isLoading={isLoading}
             state={states.purple}