+++ src/client/pages/signup/index.jsx
@@ -158,6 +158,7 @@ export default class Signup extends PureComponent {
                   <Button
                     className={styles.signup}
                     action="submit"
+                    type="submit"
                     isLoading={isLoading}
                   >
                     {'Join'}