+++ src/server/lib/next-css/css-loader-config.js
@@ -1,4 +1,4 @@
-const MiniCssExtractPlugin = require('mini-css-extract-plugin');
+const ExtractCssChunks = require('extract-css-chunks-webpack-plugin');
 const findUp = require('find-up');
 
 const fileExtensions = new Set();
@@ -33,7 +33,7 @@ module.exports = (
   }
 
   if (!isServer && !extractCssInitialized) {
-    config.plugins.push(new MiniCssExtractPlugin({
+    config.plugins.push(new ExtractCssChunks({
       // Options similar to the same options in webpackOptions.output
       // both options are optional
       filename: dev
@@ -42,6 +42,8 @@ module.exports = (
       chunkFilename: dev
         ? 'static/css/[name].chunk.css'
         : 'static/css/[name].[contenthash:8].chunk.css',
+      orderWarning: false,
+      reloadAll: true,
     }));
     extractCssInitialized = true;
   }
@@ -73,6 +75,7 @@ module.exports = (
       {},
       {
         modules: cssModules,
+        minimize: !dev,
         sourceMap: dev,
         importLoaders: loaders.length + (postcssLoader ? 1 : 0),
       },
@@ -92,7 +95,7 @@ module.exports = (
 
   return [
     !isServer && dev && 'extracted-loader',
-    !isServer && MiniCssExtractPlugin.loader,
+    !isServer && ExtractCssChunks.loader,
     cssLoader,
     postcssLoader,
     ...loaders,