+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## v0.2.0 (2018-10-26)
+
+- Fix Button component
+- Fix Error component
+- Fix css hrm
+- Update npm dependencies
+
 ## v0.1.1 (2018-10-22)
 
 - Fix eslint errors